+++ mycli/sqlexecute.py
@@ -4,9 +4,15 @@ import sqlparse
 from .packages import special
 from pymysql.constants import FIELD_TYPE
 from pymysql.converters import (convert_mysql_timestamp, convert_datetime,
-                                convert_timedelta, convert_date, conversions)
+                                convert_timedelta, convert_date, conversions,
+                                decoders)
 
 _logger = logging.getLogger(__name__)
+
+FIELD_TYPES = decoders.copy()
+FIELD_TYPES.update({
+    FIELD_TYPE.NULL: type(None)
+})
 
 class SQLExecute(object):
 
@@ -17,6 +23,7 @@ class SQLExecute(object):
     version_query = '''SELECT @@VERSION'''
 
     version_comment_query = '''SELECT @@VERSION_COMMENT'''
+    version_comment_query_mysql4 = '''SHOW VARIABLES LIKE "version_comment"'''
 
     show_candidates_query = '''SELECT name from mysql.help_topic WHERE name like "SHOW %"'''
 
@@ -111,7 +118,6 @@ class SQLExecute(object):
         # want to save them all together.
         if statement.startswith('\\fs'):
             components = [statement]
-
         else:
             components = sqlparse.split(statement)
 
@@ -221,9 +227,19 @@ class SQLExecute(object):
             _logger.debug('Version Query. sql: %r', self.version_query)
             cur.execute(self.version_query)
             version = cur.fetchone()[0]
-            _logger.debug('Version Comment. sql: %r', self.version_comment_query)
-            cur.execute(self.version_comment_query)
-            version_comment = cur.fetchone()[0].lower()
+            if version[0] == '4':
+                _logger.debug('Version Comment. sql: %r',
+                              self.version_comment_query_mysql4)
+                cur.execute(self.version_comment_query_mysql4)
+                version_comment = cur.fetchone()[1].lower()
+                if isinstance(version_comment, bytes):
+                    # with python3 this query returns bytes
+                    version_comment = version_comment.decode('utf-8')
+            else:
+                _logger.debug('Version Comment. sql: %r',
+                              self.version_comment_query)
+                cur.execute(self.version_comment_query)
+                version_comment = cur.fetchone()[0].lower()
 
         if 'mariadb' in version_comment:
             product_type = 'mariadb'