+++ test/features/steps/auto_vertical.py
@@ -40,7 +40,7 @@ def step_see_large_results(context):
     rows = ['{n:3}| {n}'.format(n=str(n)) for n in range(1, 50)]
     expected = ('***************************[ 1. row ]'
                 '***************************\r\n' +
-                '{}\r\n\r\n'.format('\r\n'.join(rows)))
+                '{}\r\n'.format('\r\n'.join(rows)))
 
     wrappers.expect_pager(context, expected, timeout=5)
     wrappers.expect_exact(context, '1 row in set', timeout=2)